+++ hydra-node/test/Hydra/OptionsSpec.hs
@@ -7,6 +7,7 @@ import Test.Hydra.Prelude
 
 import Control.Lens ((.~))
 import Data.Generics.Labels ()
+import Data.Version (Version, makeVersion)
 import Hydra.Cardano.Api (
   ChainPoint (..),
   NetworkId (..),
@@ -16,6 +17,7 @@ import Hydra.Cardano.Api (
  )
 import Hydra.Chain (maximumNumberOfParties)
 import Hydra.Network (Host (Host))
+import Hydra.NetworkVersions (hydraNodeVersion, parseNetworkTxIds)
 import Hydra.Options (
   BlockfrostOptions (..),
   CardanoChainConfig (..),
@@ -324,6 +326,15 @@ spec = parallel $
             { chainConfig = Cardano (defaultCardanoChainConfig & #chainBackendOptions .~ Blockfrost (BlockfrostOptions "blockfrost-project.txt"))
             }
 
+    it "parseNetworkTxIds produces list TxId" $ do
+      let networks = ["mainnet", "preview", "preprod"]
+      let versions = makeVersion . (\v -> [0, v, 0]) <$> [13 .. 21]
+      forM_ networks $ \network ->
+        forM_ versions $ \version -> do
+          case parseNetworkTxIds version network of
+            Left err -> failure $ "Failed to parse network tx ids: " <> err
+            Right txIds -> txIds `shouldSatisfy` not . null
+
     it "switches to offline mode when using --offline-head-seed and --initial-utxo" $
       mconcat
         [ ["--offline-head-seed", "0100"]
@@ -456,6 +467,11 @@ spec = parallel $
     prop "roundtrip parsing & printing" $
       forAll arbitrary canRoundtripRunOptionsAndPrettyPrinting
 
+    prop "parseNetworkTxIds works with expected versions and networks" $
+      forAll arbitrary $ \version ->
+        forAll arbitrary $ \network ->
+          propParseNetworkTxIds version network
+
 canRoundtripRunOptionsAndPrettyPrinting :: RunOptions -> Property
 canRoundtripRunOptionsAndPrettyPrinting opts =
   let args = toArgs opts
@@ -463,6 +479,17 @@ canRoundtripRunOptionsAndPrettyPrinting opts =
         case parseHydraCommandFromArgs args of
           Success cmd -> cmd === Run opts
           err -> property False & counterexample ("error : " <> show err)
+
+propParseNetworkTxIds :: Version -> String -> Property
+propParseNetworkTxIds version network = do
+  let varlidNetworks = ["mainnet", "preview", "preprod"]
+  let validVersions = (makeVersion . (\v -> [0, v, 0]) <$> [13 .. 21]) <> [hydraNodeVersion]
+  case parseNetworkTxIds version network of
+    Left err ->
+      if network `elem` varlidNetworks && version `elem` validVersions
+        then property False & counterexample ("error: " <> err)
+        else property True
+    Right txIds -> property $ not (null txIds)
 
 shouldParse :: HasCallStack => [String] -> Command -> Expectation
 shouldParse args cmd =